requirements.txt
pygame
selenium
qrcode
pyzbar
tensorflow
cvzone
matplotlib
pillow
screeninfo
pynput
numpy
youtube-dl
wikipedia
bs4
face_recognition
opencv-python
serial
